+++ crates/wiggle/src/error.rs
@@ -12,7 +12,7 @@ pub enum GuestError {
     #[error("Pointer out of bounds: {0:?}")]
     PtrOutOfBounds(Region),
     #[error("Pointer not aligned to {1}: {0:?}")]
-    PtrNotAligned(Region, u32),
+    PtrNotAligned(Region, u64),
     #[error("Pointer already borrowed: {0:?}")]
     PtrBorrowed(Region),
     #[error("Borrow checker out of handles")]
